refactor: consolidate wheel and rondel
