+++ game/src/board/rondel.ts
@@ -1,4 +1,4 @@
-import { GameStatePlaying, Rondel } from '../types'
+import { GameStatePlaying, Rondel, GameCommandConfigParams } from '../types'
 
 type TokenName = 'grain' | 'sheep' | 'clay' | 'coin' | 'wood' | 'joker' | 'peat' | 'grape' | 'stone'
 
@@ -31,3 +31,15 @@ export const introduceToken = (token: 'grape' | 'stone') =>
   })
 export const introduceGrapeToken = introduceToken('grape')
 export const introduceStoneToken = introduceToken('stone')
+
+export const armValues = ({ length, players }: GameCommandConfigParams) => {
+  if (players === 2 && length === 'short') {
+    return [0, 1, 2, 2, 3, 4, 4, 5, 6, 6, 7, 8, 10]
+  }
+  return [0, 2, 3, 4, 5, 6, 6, 7, 7, 8, 8, 9, 10]
+}
+
+export const take = (armIndex: number, tokenIndex: number, config: GameCommandConfigParams): number => {
+  const armVals = armValues(config)
+  return armVals[(armIndex - tokenIndex + armVals.length) % armVals.length]
+}
