--- game/src/board/wheel.ts
@@ -1,13 +1,0 @@
-import { GameCommandConfigParams, GameState } from '../types'
-
-export const armValues = ({ length, players }: GameCommandConfigParams) => {
-  if (players === 2 && length === 'short') {
-    return [0, 1, 2, 2, 3, 4, 4, 5, 6, 6, 7, 8, 10]
-  }
-  return [0, 2, 3, 4, 5, 6, 6, 7, 7, 8, 8, 9, 10]
-}
-
-export const take = (armIndex: number, tokenIndex: number, config: GameCommandConfigParams): number => {
-  const armVals = armValues(config)
-  return armVals[(armIndex - tokenIndex + armVals.length) % armVals.length]
-}
